Make sure the key store alias is globally unique
If a user has installed multiple apps with this SDK, Android will throw
an exception when the second app tries to use the same key pair alias.
This will crash our SDK and the App. This commit makes sure the key is
unique by including the property id in the string.